# ggml : add ggml_soft_max_ext