@ggerganov 2023-11-29
I think the CPU performance is degraded atm (did some quick tests earlier, haven't looked deeper yet). But I suspect it's something to do with the threading.