@slaren 2023-11-30
Much larger speedup now, but it is probably broken, perplexity returns only nans.

| model                          | backend    | ngl | test       |  master 1f5cd832 | PR 6b86bcff | speedup |
| ------------------------------ | ---------- | --: | ---------- | ---------------: | ----------: | ------: |
| llama 7B mostly Q4_0           | CUDA       |  99 | pp 512     |  3761.61 ± 76.47 | 3848.99 ± 68.95 | 1.023 |
| llama 7B mostly Q4_0           | CUDA       |  99 | pp 1024    |  3565.39 ± 37.27 | 3646.32 ± 39.29 | 1.022 |
| llama 7B mostly Q4_0           | CUDA       |  99 | pp 4096    |  2523.32 ± 12.70 | 3039.41 ± 23.85 | 1.204 |
| llama 7B mostly Q4_0           | CUDA       |  99 | pp 8192    |   1776.08 ± 8.35 | 2502.62 ± 15.85 | 1.409 |
| llama 7B mostly Q4_0           | CUDA       |  99 | pp 16384   |   1124.91 ± 6.93 | 1745.35 ± 16.31 | 1.551 |
| llama 7B mostly Q4_0           | CUDA       |  99 | tg 128     |    124.38 ± 0.81 |   127.22 ± 1.15 | 1.022 |