@ggerganov 2023-11-30
Thanks for checking - the shared arrays were not initialized so the warps were reducing random data. Should be fixed now